+++ lib/core/models/video/encoding/video_encoding.dart
@@ -42,8 +42,11 @@ class VideoEncoding {
   final AviEncodingConfig aviEncodingConfig;
 
   /// Returns the FFmpeg arguments for the specified [outputFormat].
-  List<String> toFFmpegArgs(VideoOutputFormat outputFormat) {
-    return getEncodingConfig(outputFormat).toFFmpegArgs();
+  List<String> toFFmpegArgs({
+    required VideoOutputFormat outputFormat,
+    required bool enableAudio,
+  }) {
+    return getEncodingConfig(outputFormat).toFFmpegArgs(enableAudio);
   }
 
   /// Returns the encoding configuration matching the given [outputFormat].
@@ -62,5 +65,47 @@ class VideoEncoding {
       case VideoOutputFormat.gif:
         return gifEncodingConfig;
     }
+  }
+
+  /// Returns a copy of this config with the given fields replaced.
+  VideoEncoding copyWith({
+    Mp4EncodingConfig? mp4EncodingConfig,
+    MovEncodingConfig? movEncodingConfig,
+    MkvEncodingConfig? mkvEncodingConfig,
+    WebMEncodingConfig? webMEncodingConfig,
+    GifEncodingConfig? gifEncodingConfig,
+    AviEncodingConfig? aviEncodingConfig,
+  }) {
+    return VideoEncoding(
+      mp4EncodingConfig: mp4EncodingConfig ?? this.mp4EncodingConfig,
+      movEncodingConfig: movEncodingConfig ?? this.movEncodingConfig,
+      mkvEncodingConfig: mkvEncodingConfig ?? this.mkvEncodingConfig,
+      webMEncodingConfig: webMEncodingConfig ?? this.webMEncodingConfig,
+      gifEncodingConfig: gifEncodingConfig ?? this.gifEncodingConfig,
+      aviEncodingConfig: aviEncodingConfig ?? this.aviEncodingConfig,
+    );
+  }
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+
+    return other is VideoEncoding &&
+        other.mp4EncodingConfig == mp4EncodingConfig &&
+        other.movEncodingConfig == movEncodingConfig &&
+        other.mkvEncodingConfig == mkvEncodingConfig &&
+        other.webMEncodingConfig == webMEncodingConfig &&
+        other.gifEncodingConfig == gifEncodingConfig &&
+        other.aviEncodingConfig == aviEncodingConfig;
+  }
+
+  @override
+  int get hashCode {
+    return mp4EncodingConfig.hashCode ^
+        movEncodingConfig.hashCode ^
+        mkvEncodingConfig.hashCode ^
+        webMEncodingConfig.hashCode ^
+        gifEncodingConfig.hashCode ^
+        aviEncodingConfig.hashCode;
   }
 }